pyramid: remove prebuilt init binary
This removes the need for the HTC init binary, the only problem left
with the normal init was that the modem firmware could not be loaded
because /vendor/firmware partition wasn't mounted. This is resolved by
mounting it in the early-init stage such that ueventd can find the
firmware blobs and pass them to the kernel.

Change-Id: I7ba4933192305f611eab054adc186ca1dfaf62e2